server/requirements.txt
flask>=2.2,<3
requests>=2.31,<3
python-dotenv>=1.0,<2